R57

+++ app/database/migrations/2021_09_09_043849_add_land_to_dominions.php
@@ -14,7 +14,7 @@ class AddLandToDominions extends Migration
     public function up()
     {
         Schema::table('dominions', function (Blueprint $table) {
-            $table->decimal('land')->default(0)->after('ruler_name');
+            $table->unsignedInteger('land')->default(0)->after('ruler_name');
         });
     }
